4769: Create migration, and move project steps into timeline entries
