+++ spec/factories/timeline_entries.rb
@@ -1,0 +1,4 @@
+FactoryGirl.define do
+  factory :timeline_entry do
+  end
+end
